datasets.py
import os
import logging
from PIL import Image
import torch
from torch.utils.data import Dataset

# Set up basic logging configuration
logging.basicConfig(level=logging.INFO, format='%(levelname)s: %(message)s')

class DomainDataset(Dataset):
    """
    DomainDataset loads images from a given directory and applies the specified transformations.
    
    For the 'source' domain, if load_annotations is True, it attempts to load annotations.
    For now, this is implemented with placeholders.
    
    For the 'target' domain, no annotations are expected and a placeholder (-1) is returned.
    Additionally, the file path is returned as metadata for further processing.
    """
    def __init__(self, root_dir, transform=None, domain='source', load_annotations=False):
        """
        Args:
            root_dir (str): Path to the directory containing images.
            transform (callable, optional): Transformations to apply to the images.
            domain (str): 'source' or 'target'.
            load_annotations (bool): Whether to load annotation data (only applicable for source).
        """
        self.root_dir = root_dir
        self.transform = transform
        self.domain = domain.lower()
        self.load_annotations = load_annotations

        # List image files (supporting jpg and png)
        self.image_files = [f for f in os.listdir(root_dir) if f.lower().endswith(('.jpg', '.png'))]
        if len(self.image_files) == 0:
            logging.warning(f"No image files found in directory {root_dir}")

    def __len__(self):
        return len(self.image_files)

    def _load_annotation(self, image_filename):
        """
        Placeholder function for loading annotations.
        In a complete detection setup, this should parse the corresponding annotation file.
        For now, it returns a dummy value (0).
        """
        # Here you would add code to load annotation data (bounding boxes, labels, etc.)
        return 0  # Dummy annotation

    def __getitem__(self, idx):
        image_filename = self.image_files[idx]
        image_path = os.path.join(self.root_dir, image_filename)
        try:
            image = Image.open(image_path).convert("RGB")
        except Exception as e:
            logging.error(f"Error loading image {image_path}: {e}")
            # Return a black image tensor and dummy values if loading fails.
            image = Image.new("RGB", (512, 512))
        
        if self.transform:
            image = self.transform(image)
        
        # For source domain, load annotation if flag is set; otherwise, use placeholder.
        if self.domain == 'source':
            label = self._load_annotation(image_filename) if self.load_annotations else 0
        else:
            label = -1  # For target domain, no annotations are available.

        # Return additional metadata: image filename (or full path) as unique identifier.
        return image, label, image_filename

# Example usage for testing (can be removed in production)
if __name__ == "__main__":
    from torchvision import transforms

    # Define a simple transformation pipeline
    transform = transforms.Compose([
        transforms.Resize((512, 512)),
        transforms.ToTensor()
    ])

    # Test with a source domain dataset (assuming images exist in 'data/source/train')
    source_dataset = DomainDataset(root_dir="data/source/train", transform=transform, domain='source', load_annotations=True)
    logging.info(f"Source dataset size: {len(source_dataset)}")
    sample_img, sample_label, sample_id = source_dataset[0]
    logging.info(f"Sample Source - Image shape: {sample_img.shape}, Label: {sample_label}, ID: {sample_id}")

    # Test with a target domain dataset (assuming images exist in 'data/target/foggy/train')
    target_dataset = DomainDataset(root_dir="data/target/foggy/train", transform=transform, domain='target')
    logging.info(f"Target dataset size: {len(target_dataset)}")
    sample_img, sample_label, sample_id = target_dataset[0]
    logging.info(f"Sample Target - Image shape: {sample_img.shape}, Label: {sample_label}, ID: {sample_id}")
